+++ mayan/apps/document_states/links.py
@@ -4,6 +4,7 @@ from django.utils.translation import ugettext_lazy as _
 
 from mayan.apps.documents.permissions import permission_document_type_edit
 from mayan.apps.navigation.classes import Link
+from mayan.apps.navigation.utils import get_cascade_condition
 
 from .permissions import (
     permission_workflow_create, permission_workflow_delete,
@@ -207,8 +208,10 @@ link_workflow_runtime_proxy_document_list = Link(
     view='document_states:workflow_runtime_proxy_document_list',
 )
 link_workflow_runtime_proxy_list = Link(
-    icon_class_path='mayan.apps.document_states.icons.icon_workflow_runtime_proxy_list',
-    permissions=(permission_workflow_view,),
+    condition=get_cascade_condition(
+        app_label='document_states', model_name='WorkflowRuntimeProxy',
+        object_permission=permission_workflow_view,
+    ), icon_class_path='mayan.apps.document_states.icons.icon_workflow_runtime_proxy_list',
     text=_('Workflows'), view='document_states:workflow_runtime_proxy_list'
 )
 link_workflow_runtime_proxy_state_document_list = Link(