+++ HISTORY.rst
@@ -100,6 +100,9 @@
   user management.
 - Split document app links.
 - Make Postgres container wait delay configurable.
+- Enable the sidebar workflow runtime link when
+  the workflow view permission is granted to at
+  least one workflow.
 
 3.2.9 (2019-XX-XX)
 ==================